Add the NES 6502 CPU emulator
This is the initial commit to the REMulator repo. This is a large commit
because quite a bit of work had been done on this prior to creating the
repo.

The bulk of this change is in implementing the 6502 CPU. It supports all
of the official opcodes, as well as many unofficial opcodes. It supports
accurate timing of each opcode: including extra cycle costs due to page
boundary crossing, successful branching, etc.

This has been tested against a comprehensive test ROM designed to test
the NES CPU.